Set up umbrella headers
To trigger early warnings/errors.
