+++ Sources/PartoutCore/Connection/ReachabilityInfo.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-internal import _PartoutCore_C
+internal import _PartoutPortable_C
 
 /// Platform-specific metadata about network reachability.
 public struct ReachabilityInfo: Sendable {
