+++ Sources/PartoutCore/Connection/TunInterface.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-internal import _PartoutCore_C
+internal import _PartoutPortable_C
 
 /// Represents a specific I/O interface meant to work at a L3 virtual tun layer.
 public typealias TunInterface = IOInterface
