+++ Sources/PartoutCore/Data/CZeroingData.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-internal import _PartoutCore_C
+internal import _PartoutPortable_C
 
 /// Wrapper of binary data that zeroes out itself on deallocation.
 public final class CZeroingData {
