+++ Sources/PartoutCore/Data/Data+Network.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-internal import _PartoutCore_C
+internal import _PartoutPortable_C
 
 extension Data {
     public var asIPAddress: String? {
