+++ Sources/PartoutCore/Data/Data+Zeroing.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-internal import _PartoutCore_C
+internal import _PartoutPortable_C
 
 extension Data {
     public static func zeroing(_ zd: UnsafeMutableRawPointer) -> Self {
