+++ Sources/PartoutCore/OpenAPI/Core/SecureData.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-internal import _PartoutCore_C
+internal import _PartoutPortable_C
 
 /// Wrapper of a byte array with safe encoding capabilities.
 public struct SecureData: Hashable, Codable, @unchecked Sendable {
