+++ Sources/PartoutCore/OpenAPI/Core/Subnet.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-internal import _PartoutCore_C
+internal import _PartoutPortable_C
 
 /// An IPv4/v6 subnet.
 public struct Subnet: Hashable, Codable, Sendable {
