+++ Sources/PartoutCore/PRNG/PlatformPRNG.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-internal import _PartoutCore_C
+internal import _PartoutPortable_C
 
 /// Implementation of ``PRNGProtocol`` with the OS C library from MiniFoundation.
 public final class PlatformPRNG: PRNGProtocol {
