+++ Sources/PartoutCore_C/include/crypto/crypto.h
@@ -6,180 +6,9 @@
 
 #pragma once
 
-#include "portable/common.h"
-#include "portable/zd.h"
-
-#pragma clang assume_nonnull begin
-
-/// - Parameters:
-///   - size: The base number of bytes.
-///   - overhead: The extra number of bytes.
-/// - Returns: The number of bytes to store a crypto buffer safely.
-static inline
-size_t pp_crypto_encryption_base_capacity(size_t size, size_t overhead) {
-    // AES only, block is 128-bit
-    static const int PP_CRYPTO_MAX_BLOCK_SIZE = 16;
-    // Encryption, byte-alignment, overhead (e.g. IV, digest)
-    return 2 * size + PP_CRYPTO_MAX_BLOCK_SIZE + overhead;
-}
-
-bool pp_crypto_init_seed(const uint8_t *src, const size_t len);
-
-typedef enum {
-    PPCryptoErrorNone,
-    PPCryptoErrorEncryption,
-    PPCryptoErrorHMAC
-} pp_crypto_error_code;
-
-typedef struct {
-    const pp_zd *enc_key;
-    const pp_zd *dec_key;
-} pp_crypto_key_pair;
-
-typedef struct {
-    pp_crypto_key_pair cipher;
-    pp_crypto_key_pair hmac;
-} pp_crypto_keys;
-
-/// Custom flags for encryption routines.
-typedef struct {
-    /// A custom initialization vector (IV).
-    const uint8_t *_Nullable iv;
-
-    /// The length of ``iv``.
-    size_t iv_len;
-
-    /// A custom associated data for AEAD (AD).
-    const uint8_t *_Nullable ad;
-
-    /// The length of ``ad``.
-    size_t ad_len;
-
-    /// Enable testable (predictable) behavior.
-    int for_testing;
-} pp_crypto_flags;
-
-typedef void (*pp_crypto_configure_fn)(void *ctx,
-                                       const pp_zd *_Nullable cipher_key,
-                                       const pp_zd *_Nullable hmac_key);
-
-typedef size_t (*pp_crypto_encrypt_fn)(void *ctx,
-                                       uint8_t *out, size_t out_buf_len,
-                                       const uint8_t *in, size_t in_len,
-                                       const pp_crypto_flags *_Nullable flags,
-                                       pp_crypto_error_code *_Nullable error);
-
-typedef size_t (*pp_crypto_decrypt_fn)(void *ctx,
-                                       uint8_t *out, size_t out_buf_len,
-                                       const uint8_t *in, size_t in_len,
-                                       const pp_crypto_flags *_Nullable flags,
-                                       pp_crypto_error_code *_Nullable error);
-
-typedef bool (*pp_crypto_verify_fn)(void *ctx,
-                                    const uint8_t *in, size_t in_len,
-                                    pp_crypto_error_code *_Nullable error);
-
-typedef struct {
-    pp_crypto_configure_fn configure;
-    pp_crypto_encrypt_fn encrypt;
-} pp_crypto_encrypter;
-
-typedef struct {
-    pp_crypto_configure_fn configure;
-    pp_crypto_decrypt_fn decrypt;
-    pp_crypto_verify_fn verify;
-} pp_crypto_decrypter;
-
-typedef size_t (*pp_crypto_capacity_fn)(const void *ctx, size_t len);
-
-typedef struct {
-    size_t cipher_key_len;
-    size_t cipher_iv_len;
-    size_t hmac_key_len;
-    size_t digest_len;
-    size_t tag_len;
-    pp_crypto_capacity_fn encryption_capacity;
-} pp_crypto_meta;
-
-typedef struct {
-    pp_crypto_meta meta;
-    pp_crypto_encrypter encrypter;
-    pp_crypto_decrypter decrypter;
-} pp_crypto;
-
-typedef struct {
-    pp_crypto base;
-} *pp_crypto_ctx;
-
-typedef void (*pp_crypto_free_fn)(pp_crypto_ctx);
-
-#ifndef MAX
-#define MAX(a,b) ((a) > (b) ? (a) : (b))
-#endif
-
-#ifndef MIN
-#define MIN(a,b) ((a) < (b) ? (a) : (b))
-#endif
-
-static inline
-size_t pp_crypto_encryption_capacity(pp_crypto_ctx ctx, size_t len) {
-    return ctx->base.meta.encryption_capacity(&ctx->base, len);
-}
-
-static inline
-void pp_crypto_configure_encrypt(pp_crypto_ctx ctx,
-                                 const pp_zd *_Nullable cipher_key,
-                                 const pp_zd *_Nullable hmac_key) {
-    ctx->base.encrypter.configure(&ctx->base, cipher_key, hmac_key);
-}
-
-static inline
-size_t pp_crypto_encrypt(pp_crypto_ctx ctx,
-                         uint8_t *out, size_t out_buf_len,
-                         const uint8_t *in, size_t in_len,
-                         const pp_crypto_flags *_Nullable flags,
-                         pp_crypto_error_code *_Nullable error) {
-    return ctx->base.encrypter.encrypt(&ctx->base, out, out_buf_len, in, in_len, flags, error);
-}
-
-static inline
-void pp_crypto_configure_decrypt(pp_crypto_ctx ctx,
-                                 const pp_zd *_Nullable cipher_key,
-                                 const pp_zd *_Nullable hmac_key) {
-    ctx->base.decrypter.configure(&ctx->base, cipher_key, hmac_key);
-}
-
-static inline
-size_t pp_crypto_decrypt(pp_crypto_ctx ctx,
-                         uint8_t *out, size_t out_buf_len,
-                         const uint8_t *in, size_t in_len,
-                         const pp_crypto_flags *_Nullable flags,
-                         pp_crypto_error_code *_Nullable error) {
-    return ctx->base.decrypter.decrypt(&ctx->base, out, out_buf_len, in, in_len, flags, error);
-}
-
-static inline
-bool pp_crypto_verify(pp_crypto_ctx ctx,
-                      const uint8_t *in, size_t in_len,
-                      pp_crypto_error_code *_Nullable error) {
-    pp_assert(ctx->base.decrypter.verify);
-    return ctx->base.decrypter.verify(&ctx->base, in, in_len, error);
-}
-
-static inline
-pp_crypto_meta pp_crypto_meta_of(pp_crypto_ctx ctx) {
-    return ctx->base.meta;
-}
-
-static inline
-void pp_assert_encryption_length(size_t out_len, size_t in_len) {
-    const size_t out_min_len = pp_crypto_encryption_base_capacity(in_len, 0);
-    pp_assert(out_len >= out_min_len);
-}
-
-static inline
-void pp_assert_decryption_length(size_t out_len, size_t in_len) {
-    pp_assert(out_len >= in_len);
-}
-
-#pragma clang assume_nonnull end
+#include "crypto.h"
+#include "crypto_aead.h"
+#include "crypto_cbc.h"
+#include "crypto_ctr.h"
+#include "hmac.h"
+#include "keys.h"
