+++ Sources/PartoutCore_C/include/portable/portable.h
@@ -1,0 +1,16 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Davide De Rosa
+ *
+ * SPDX-License-Identifier: GPL-3.0
+ */
+
+#pragma once
+
+#include "common.h"
+#include "endian.h"
+#include "mux.h"
+#include "network.h"
+#include "prng.h"
+#include "socket.h"
+#include "tun.h"
+#include "zd.h"
