+++ Sources/PartoutOpenVPN_C/include/openvpn/openvpn.h
@@ -1,0 +1,23 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Davide De Rosa
+ *
+ * SPDX-License-Identifier: GPL-3.0
+ */
+
+#include "openvpn/comp.h"
+#include "openvpn/control.h"
+#include "openvpn/dp_error.h"
+#include "openvpn/dp_framing.h"
+#include "openvpn/dp_framing_comp.h"
+#include "openvpn/dp_macros.h"
+#include "openvpn/dp_mode.h"
+#include "openvpn/dp_mode_ad.h"
+#include "openvpn/dp_mode_hmac.h"
+#include "openvpn/dp_mode_shortcuts.h"
+#include "openvpn/mss_fix.h"
+#include "openvpn/obf.h"
+#include "openvpn/packet.h"
+#include "openvpn/pkt_proc.h"
+#include "openvpn/replay.h"
+#include "openvpn/test/crypto_mock.h"
+#include "openvpn/test/dp_mock.h"
