+++ Sources/Partout_C/partout.c
@@ -3,3 +3,11 @@
  *
  * SPDX-License-Identifier: GPL-3.0
  */
+
+#if PARTOUT_OPENVPN
+#include "openvpn/openvpn.h"
+#endif
+#if PARTOUT_WIREGUARD
+#include "wireguard/wireguard.h"
+#include "wireguard/backend.h"
+#endif
